Smart_search/requirements.txt
fastapi
uvicorn
torch
torchvision
scikit-learn
pillow
